+++ src/main/java/roomescape/config/WebMvcConfiguration.java
@@ -5,8 +5,8 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.argumentresolver.CheckAdminInterceptor;
-import roomescape.argumentresolver.TokenValueMethodArgumentResolver;
+import roomescape.config.Interceptor.CheckAdminInterceptor;
+import roomescape.config.resolver.TokenValueMethodArgumentResolver;
 
 @Configuration
 public class WebMvcConfiguration implements WebMvcConfigurer {
