--- src/main/java/roomescape/service/AdminReservationService.java
@@ -1,68 +1,0 @@
-package roomescape.service;
-
-import java.util.NoSuchElementException;
-import org.springframework.stereotype.Service;
-import roomescape.domain.Member;
-import roomescape.domain.Reservation;
-import roomescape.domain.ReservationTime;
-import roomescape.domain.Theme;
-import roomescape.persistence.MemberRepository;
-import roomescape.persistence.ReservationRepository;
-import roomescape.persistence.ReservationTimeRepository;
-import roomescape.persistence.ThemeRepository;
-import roomescape.service.request.AdminReservationRequest;
-import roomescape.service.response.ReservationResponse;
-
-@Service
-public class AdminReservationService {
-    private final ReservationRepository reservationRepository;
-    private final MemberRepository memberRepository;
-    private final ReservationTimeRepository reservationTimeRepository;
-    private final ThemeRepository themeRepository;
-
-    public AdminReservationService(
-            ReservationRepository reservationRepository,
-            MemberRepository memberRepository,
-            ReservationTimeRepository reservationTimeRepository,
-            ThemeRepository themeRepository
-    ) {
-        this.reservationRepository = reservationRepository;
-        this.memberRepository = memberRepository;
-        this.reservationTimeRepository = reservationTimeRepository;
-        this.themeRepository = themeRepository;
-    }
-
-    public ReservationResponse createReservation(AdminReservationRequest request) {
-        Member member = getMember(request.memberId());
-        ReservationTime reservationTime = getReservationTime(request.timeId());
-        Theme theme = getTheme(request.themeId());
-        Reservation reservation = request.toDomain(member, reservationTime, theme);
-        reservation.validateDateTime();
-        validateDuplicateReservation(reservation);
-
-        Reservation createdReservation = reservationRepository.create(reservation);
-        return ReservationResponse.from(createdReservation);
-    }
-
-    private Member getMember(Long id) {
-        System.out.println("id = " + id);
-        return memberRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException("해당되는 회원이 없습니다."));
-    }
-
-    private ReservationTime getReservationTime(Long id) {
-        return reservationTimeRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException("해당되는 예약 시간이 없습니다."));
-    }
-
-    private Theme getTheme(Long id) {
-        return themeRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException("해당되는 테마가 없습니다."));
-    }
-
-    private void validateDuplicateReservation(Reservation reservation) {
-        if (reservationRepository.hasDuplicateReservation(reservation)) {
-            throw new IllegalStateException("중복된 예약이 존재합니다.");
-        }
-    }
-}
