+++ src/main/java/roomescape/web/controller/AdminReservationController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.service.AdminReservationService;
+import roomescape.service.ReservationService;
 import roomescape.service.request.AdminReservationRequest;
 import roomescape.service.response.ReservationResponse;
 
@@ -14,15 +14,15 @@ import roomescape.service.response.ReservationResponse;
 @RequestMapping("/admin")
 public class AdminReservationController {
 
-    private final AdminReservationService reservationService;
+    private final ReservationService reservationService;
 
-    public AdminReservationController(AdminReservationService reservationService) {
+    public AdminReservationController(ReservationService reservationService) {
         this.reservationService = reservationService;
     }
 
     @PostMapping("/reservations")
     public ResponseEntity<ReservationResponse> createReservation(@RequestBody AdminReservationRequest request) {
-        ReservationResponse response = reservationService.createReservation(request);
+        ReservationResponse response = reservationService.createAdminReservation((request));
         URI uri = URI.create("/reservations/" + response.id());
 
         return ResponseEntity.created(uri).body(response);
