+++ src/main/java/roomescape/web/controller/AuthorizationPageController.java
@@ -4,7 +4,7 @@ import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class LoginPageController {
+public class AuthorizationPageController {
 
     @GetMapping("/login")
     public String getLoginPage() {
